requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
kafka-python==2.0.2
requests==2.28.1
psycopg2-binary==2.9.5
